seperated getInspectedResources and Parser
